refactor(docs): abstract scrollToAnchor
